+++ programs/hook_program/src/instructions/mod.rs
@@ -1,5 +1,5 @@
 pub mod add_associated_protocol;
-pub mod deposit_hook;
+pub mod execute_deposit_hook;
 pub mod get_nav_data;
 pub mod init_vault_associated_protocols;
 pub mod initialize_deposit_extra_meta_accounts;
@@ -7,7 +7,7 @@ pub mod remove_associated_protocol;
 pub mod update_nav_data;
 
 pub use add_associated_protocol::*;
-pub use deposit_hook::*;
+pub use execute_deposit_hook::*;
 pub use get_nav_data::*;
 pub use init_vault_associated_protocols::*;
 pub use initialize_deposit_extra_meta_accounts::*;
