+++ programs/vault/src/extensions/deposit_hook.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+#[derive(AnchorDeserialize, AnchorSerialize, Clone, InitSpace, Copy)]
+pub struct DepositHook {
+    pub hook_program_id: Pubkey,
+    pub authority: Pubkey,
+}
