+++ programs/vault/src/extensions/mod.rs
@@ -1,3 +1,7 @@
+pub mod deposit_hook;
 pub mod extra_meta_accounts;
+pub mod vault_extension;
 
+pub use deposit_hook::*;
 pub use extra_meta_accounts::*;
+pub use vault_extension::*;
