+++ programs/vault/src/state/mod.rs
@@ -1,5 +1,7 @@
 pub mod constants;
+pub mod fee;
 pub mod vault_config;
 
 pub use constants::*;
+pub use fee::*;
 pub use vault_config::*;
